Move planes.tsx into Planes folder
